+++ src/lib/modules/Landing/EggBox/index.tsx
@@ -1,0 +1,1 @@
+export { default } from "./EggBox";
